requirements.txt
# Core Django dependencies
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.5

# Database
dj-database-url==3.0.1
psycopg2-binary==2.9.10

# Supabase and related
supabase==2.17.0
gotrue==2.12.3
httpx==0.28.1
storage3==0.12.0
realtime==2.0.0

# Channels for WebSocket
channels==4.2.2
channels_redis==4.2.1
websockets==11.0.3

# File handling
pillow==11.2.1
django-storages==1.14.2

# Payment
razorpay==1.4.2

# Google OAuth
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Other utilities
gunicorn==23.0.0
whitenoise==6.9.0
redis==6.2.0
python-decouple